historico candidato agora aceita indiferente, no campo apto

+++ src/com/fortes/rh/dao/hibernate/desenvolvimento/ColaboradorTurmaDaoHibernate.java
@@ -1,5 +1,6 @@
 package com.fortes.rh.dao.hibernate.desenvolvimento;
 
+import java.math.BigDecimal;
 import java.math.BigInteger;
 import java.util.ArrayList;
 import java.util.Collection;
@@ -909,7 +910,7 @@ public class ColaboradorTurmaDaoHibernate extends GenericDaoHibernate<Colaborado
 			
 			ct.setCurso(new Curso());
 			if(res[3] != null)
-				ct.getCurso().setPercentualMinimoFrequencia((Double)res[3]);
+				ct.getCurso().setPercentualMinimoFrequencia(((BigDecimal)res[3]).doubleValue());
 			if(res[4] != null)
 				ct.setQtdAvaliacoesCurso(((BigInteger)res[4]).intValue());
 			if(res[5] != null)
